(split)bug: 38768
Missed a few spots where we truncated the mailbox ID or the group ID to an int.

http://bugzilla.zimbra.com/show_bug.cgi?id=38768

Copied from Perforce
 Change: 165880

+++ src/java/com/zimbra/cs/db/DbMailbox.java
@@ -695,7 +695,7 @@ public class DbMailbox {
             Mailbox.MailboxData mbd = new Mailbox.MailboxData();
             mbd.id            = mailboxId;
             mbd.accountId     = rs.getString(pos++);
-            mbd.schemaGroupId = rs.getInt(pos++);
+            mbd.schemaGroupId = rs.getLong(pos++);
             mbd.size          = rs.getLong(pos++);
             if (rs.wasNull())
                 mbd.size = -1;
